fix: 가게, 메뉴, 한줄리뷰, 검색 서비스/구현체 분리
